pyproject.toml
[tool.poetry]
name = "miditest"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.6,<3.11"
python-rtmidi = "^1.4.9"
imgui = {git = "https://github.com/pyimgui/pyimgui.git", branch = "dev/version-2.0", extras = ["glfw"]}
ptimer = "^0.1.1"
dataclasses-json = "^0.5.7"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
